Getting Started ===============
Here we simply setup a new rails application and scaffold a
new model to hold our concept of a planned meal.

I've opted to use the new importmap Javascript style, which is the
default and to use TailwindCSS for CSS, but these are personal
preference only. It's also the default to setup Turbo and Stimulus when
creating a new project.

```bash
$ rails new hotwire-demo-meal-planner --database=postgresql --css=tailwind
```

I've introduced `docker-compose.yml` file to run the database and redis
services so I don't have to manage their installations on my machine.

The `Meal` model is simply the name of the meal and the date on which it
was planned.

```bash
$ bin/rails generate scaffold meal meal:string date:date
```

I've then cleaned up some things I don't need from the generators,
including the `show` action for `meals` - the meals will be shown in
their entirity as part of the `index` action for the sake of this demo.