Plan meals in place
Leverage turbo frames to allow the user to plan meals in-place.

We need to redesign the meal form to look more at home within the
calendar list, but once that's done the only job is to wrap the shown
meal and the meal form in a turbo-frame-tag that have the same ID. Turbo
does the rest.

To make the design changes, we've made a solid change to functionality
of the form - we no longer allow users to select a date from a date
picker, instead the date is set once the user selects to plan a meal.

This means we've lost the ability to 'move' a planned meal.