Enable real time updates
Subscribe to changes to meals to update the calendar in real time. This
allows multiple users to collborate on planning their meals and see each
others changes in real time.

We can't make use of the normal helper since creating and deleting meals
does not cause an insert or removal from the calendar list. Instead
every change to meals causes a replace on the associated date.